EXYNOS: change the compile/clean option of spl
spl tool of exynos5 designed only for smdk5250.
It compiled when defined CONFIG_SMDK5250 and BOARD name is smdk5250.
It means that it's not considered shrink boards with different board name.

So, it changed compile condition to exynos5 family option
not a specific board name for cover the other board.
I also modified clean condition to clear compiled spl.

Otherwise, tools/mk*spl show on untracked files.
It should be ignore in git.

Signed-off-by: Jeong-Hyeon Kim <jhkim@insignal.co.kr>