Exynos5: Fix errata 773022 and 774769 on Exynos5250
The exynos5250 chip is affected by the errata

  773022 (disable loop buffer)
  774769 (refrain streaming-write to ever allocate into the L2 cache)

and needs to get the respective fixes as mentioned above applied.

This is done in the Samsung downstream kernel within Linux. However, that
approach fails to work when we run the guest kernel in HYP mode, as in
now Linux's errata fixup code runs without privileges to access the
aux control register.

So instead, let's patch up u-boot (where CPU errata fixes belong to
anyways), so that we can happily run a kernel.

This fixes random segmentation faults on Arndale for me, that mostly
occured because of the broken loop buffer.

Signed-off-by: Alexander Graf <agraf@suse.de>